refresh dir structure
