+++ packages/parsers/src/openapi/3.1/__test__/paths/OperationObjectConverter.node.test.ts
@@ -2,7 +2,7 @@ import { FdrAPI } from "@fern-api/fdr-sdk";
 import { OpenAPIV3_1 } from "openapi-types";
 import { describe, expect, it } from "vitest";
 import { createMockContext } from "../../../../__test__/createMockContext.util";
-import { OperationObjectConverterNode } from "../OperationObjectConverter.node";
+import { OperationObjectConverterNode } from "../../paths/OperationObjectConverter.node";
 
 describe("OperationObjectConverterNode", () => {
     const mockContext = createMockContext();
