+++ packages/parsers/src/openapi/3.1/__test__/paths/PathItemObjectConverter.node.test.ts
@@ -1,7 +1,7 @@
 import { OpenAPIV3_1 } from "openapi-types";
 import { describe, expect, it } from "vitest";
 import { createMockContext } from "../../../../__test__/createMockContext.util";
-import { PathItemObjectConverterNode } from "../PathItemObjectConverter.node";
+import { PathItemObjectConverterNode } from "../../paths/PathItemObjectConverter.node";
 
 describe("PathItemObjectConverterNode", () => {
     const mockContext = createMockContext();
