+++ packages/parsers/src/openapi/3.1/__test__/schemas/ObjectConverter.node.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { createMockContext } from "../../../../__test__/createMockContext.util";
-import { ObjectConverterNode } from "../ObjectConverter.node";
-import { SchemaConverterNode } from "../SchemaConverter.node";
+import { ObjectConverterNode } from "../../schemas/ObjectConverter.node";
+import { SchemaConverterNode } from "../../schemas/SchemaConverter.node";
 
 describe("ObjectConverterNode", () => {
     const mockContext = createMockContext();
