+++ packages/parsers/src/openapi/3.1/__test__/schemas/OneOfConverter.node.test.ts
@@ -1,5 +1,5 @@
 import { createMockContext } from "../../../../__test__/createMockContext.util";
-import { OneOfConverterNode } from "../OneOfConverter.node";
+import { OneOfConverterNode } from "../../schemas/OneOfConverter.node";
 
 describe("OneOfConverterNode", () => {
     const mockContext = createMockContext({
