+++ packages/parsers/src/openapi/3.1/__test__/schemas/ReferenceConverter.node.test.ts
@@ -1,7 +1,7 @@
 import { OpenAPIV3_1 } from "openapi-types";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { createMockContext } from "../../../../__test__/createMockContext.util";
-import { ReferenceConverterNode } from "../ReferenceConverter.node";
+import { ReferenceConverterNode } from "../../schemas/ReferenceConverter.node";
 
 describe("ReferenceConverterNode", () => {
     const mockContext = createMockContext();
