+++ packages/parsers/src/openapi/3.1/__test__/schemas/primitives/EnumConverter.node.test.ts
@@ -1,7 +1,7 @@
 import { OpenAPIV3_1 } from "openapi-types";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { createMockContext } from "../../../../__test__/createMockContext.util";
-import { EnumConverterNode } from "../EnumConverter.node";
+import { createMockContext } from "../../../../../__test__/createMockContext.util";
+import { EnumConverterNode } from "../../../schemas/primitives/EnumConverter.node";
 
 describe("EnumConverterNode", () => {
     const mockContext = createMockContext();
