+++ packages/parsers/src/openapi/3.1/paths/PathItemObjectConverter.node.ts
