+++ packages/parsers/src/openapi/3.1/paths/PathsObjectConverter.node.ts
