+++ packages/parsers/src/openapi/3.1/schemas/primitives/NullConverter.node.ts
@@ -1,6 +1,6 @@
 import { FdrAPI } from "@fern-api/fdr-sdk";
 import { OpenAPIV3_1 } from "openapi-types";
-import { BaseOpenApiV3_1Node, BaseOpenApiV3_1NodeConstructorArgs } from "../../BaseOpenApiV3_1Converter.node";
+import { BaseOpenApiV3_1Node, BaseOpenApiV3_1NodeConstructorArgs } from "../../../BaseOpenApiV3_1Converter.node";
 
 export declare namespace NullConverterNode {
     export interface Input extends OpenAPIV3_1.NonArraySchemaObject {
